splits up models and serializers into small files
