+++ server/pos/models/stock.py
@@ -1,13 +1,6 @@
 from django.db import models
 
-
-class User(models.Model):
-    name = models.CharField(max_length=255)
-    max_credit = models.IntegerField()
-    card = models.CharField(max_length=255)
-
-    def __str__(self):
-        return self.name
+from .user import User
 
 
 class Ingredient(models.Model):
@@ -84,12 +77,3 @@ class Purchase:
             s += '\n' + str(line)
 
         return s
-
-
-class Shift(models.Model):
-    start = models.DateTimeField(auto_now_add=True, blank=False)
-    end = models.DateTimeField(blank=True, null=True)
-    leader = models.ForeignKey(User)
-
-    def __str__(self):
-        return 'Skift ledet av ' + str(self.leader) + ' som startet ' + self.start.strftime('%Y-%m-%d %H:%M:%S')
