+++ server/pos/serializers/shift.py
@@ -1,0 +1,9 @@
+from pos.models.shift import Shift
+
+from rest_framework import serializers
+
+
+class ShiftSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Shift
+        fields = ('id', 'start', 'end', 'leader')
